+++ src/stores/draggable/DraggableStore.js
@@ -1,0 +1,14 @@
+/**
+ * @typedef {ReturnType<createStore>} Store
+ */
+
+export function createStore() {
+  return {
+    dragTargetId: '',
+    dragOverTargetId: '',
+    dragStart: [0, 0],
+    dragMove: [0, 0],
+    dragStop: [0, 0],
+    dragging: false,
+  };
+}